Refactor LED/Horn Annunciation
Refactored LED/Horn for much greater granularity such that if an axis is good, keep the green LED illuminated, the orange LED illuminates when an axis is in a caution context and, as originally implemented, the red LED and warning horn annunciate if an axis is in a warning context.  This will likely change when my 5mm/8mm APA106/NeoPixel LEDs arrive (desire to migrate away from discrete LEDs to a single visual indicator device, easing wiring and power drama (no resistors, multi-spectral colors, etc.).  Also, I am considering migrating from an active fixed-tone aural device to a passive PWM-controlled aural warning device, as it is moar gucci, ya know...