Integrate DJI O3 RC ARMing app
Given I developed the Xiao Inclinometer for use on a Rock Crawler that would have a DJI O3 HD FPV Camera thereon, given that the O3 Air Unit requires arming to enable the highest power and bit rate outputs, and given that both the inclinometer and the arming app are intended to be run on a Xiao, it makes perfect sense to use a single Xiao to run both.  I have folded both apps together herein.